Multiple changes to deploy scripts

+++ deploy/openshift/ocp.sh
@@ -20,12 +20,10 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
     DEFAULT_OC_PUBLIC_HOSTNAME="$LOCAL_IP_ADDRESS"
     DEFAULT_OC_PUBLIC_IP="$LOCAL_IP_ADDRESS"
     DEFAULT_OC_BINARY_DOWNLOAD_URL="https://github.com/openshift/origin/releases/download/v3.9.0/openshift-origin-client-tools-v3.9.0-191fece-mac.zip"
-    DEFAULT_JQ_BINARY_DOWNLOAD_URL="https://github.com/stedolan/jq/releases/download/jq-1.5/jq-osx-amd64"
 else
     DEFAULT_OC_PUBLIC_HOSTNAME="$LOCAL_IP_ADDRESS"
     DEFAULT_OC_PUBLIC_IP="$LOCAL_IP_ADDRESS"
     DEFAULT_OC_BINARY_DOWNLOAD_URL="https://github.com/openshift/origin/releases/download/v3.9.0/openshift-origin-client-tools-v3.9.0-191fece-linux-64bit.tar.gz"
-    DEFAULT_JQ_BINARY_DOWNLOAD_URL="https://github.com/stedolan/jq/releases/download/jq-1.5/jq-linux64"
 fi
 
 export OC_PUBLIC_HOSTNAME=${OC_PUBLIC_HOSTNAME:-${DEFAULT_OC_PUBLIC_HOSTNAME}}
@@ -35,8 +33,6 @@ DEFAULT_OCP_TOOLS_DIR="/tmp"
 export OCP_TOOLS_DIR=${OCP_TOOLS_DIR:-${DEFAULT_OCP_TOOLS_DIR}}
 
 export OC_BINARY_DOWNLOAD_URL=${OC_BINARY_DOWNLOAD_URL:-${DEFAULT_OC_BINARY_DOWNLOAD_URL}}
-export JQ_BINARY_DOWNLOAD_URL=${JQ_BINARY_DOWNLOAD_URL:-${DEFAULT_JQ_BINARY_DOWNLOAD_URL}}
-
 
 DEFAULT_OPENSHIFT_USERNAME="developer"
 export OPENSHIFT_USERNAME=${OPENSHIFT_USERNAME:-${DEFAULT_OPENSHIFT_USERNAME}}
@@ -109,7 +105,6 @@ test_dns_provider() {
 
 get_tools() {
     OC_BINARY="$OCP_TOOLS_DIR/oc"
-    JQ_BINARY="$OCP_TOOLS_DIR/jq"
     OC_VERSION=$(echo $OC_BINARY_DOWNLOAD_URL | cut -d '/' -f 8)
     #OS specific extract archives
     if [[ "$OSTYPE" == "darwin"* ]]; then
@@ -129,51 +124,20 @@ get_tools() {
         rm -f "$OCP_TOOLS_DIR"/README.md "$OCP_TOOLS_DIR"/LICENSE "${OCP_TOOLS_DIR:-/tmp}"/"$OC_PACKAGE"
     }
 
-    if [[ ! -f $OC_BINARY ]]; then
-        download_oc
-    else
-        # here we check is installed version is same version defined in script, if not we update version to one that defined in script.
-        if [[ $($OC_BINARY version 2> /dev/null | grep "oc v" | cut -d " " -f2 | cut -d '+' -f1 || true) != *"$OC_VERSION"* ]]; then
-            rm -f "$OC_BINARY" "$OCP_TOOLS_DIR"/README.md "$OCP_TOOLS_DIR"/LICENSE
-            download_oc
-        fi
+    if [[ $(oc version 2> /dev/null | grep "oc v" | cut -d " " -f2 | cut -d '+' -f1 || true) == *"$OC_VERSION"* ]]; then
+       echo "Found oc ${OC_VERSION} in PATH. Using it"
+       export OC_BINARY="oc"
+     elif [[ ! -f $OC_BINARY ]]; then
+         echo two
+          download_oc
+       else
+          # here we check is installed version is same version defined in script, if not we update version to one that defined in script.
+           if [[ $($OC_BINARY version 2> /dev/null | grep "oc v" | cut -d " " -f2 | cut -d '+' -f1 || true) != *"$OC_VERSION"* ]]; then
+               rm -f "$OC_BINARY" "$OCP_TOOLS_DIR"/README.md "$OCP_TOOLS_DIR"/LICENSE
+               download_oc
+           fi
     fi
-
-    if [ ! -f $JQ_BINARY ]; then
-        echo "download jq..."
-        wget -O $JQ_BINARY $JQ_BINARY_DOWNLOAD_URL
-        chmod +x $JQ_BINARY
-    fi
-    export PATH=${PATH}:${OCP_TOOLS_DIR}
-}
-
-ocp_is_booted() {
-    # we have to wait before docker registry will be started as it is staring as last container and it should be running before we perform che deploy.
-    ocp_registry_container_id=$(docker ps | grep k8s_registry_docker-registry | cut -d ' ' -f1)
-    if [ ! -z "$ocp_registry_container_id" ];then
-        ocp_registry_container_status=$(docker inspect "$ocp_registry_container_id" | $JQ_BINARY .[0] | $JQ_BINARY -r '.State.Status')
-    else
-        return 1
-    fi
-    if [[ "${ocp_registry_container_status}" == "running" ]]; then
-        return 0
-    else
-        return 1
-    fi
-}
-
-wait_ocp() {
-  OCP_BOOT_TIMEOUT=120
-  echo "[OCP] wait for ocp full boot..."
-  ELAPSED=0
-  until ocp_is_booted; do
-    if [ ${ELAPSED} -eq "${OCP_BOOT_TIMEOUT}" ];then
-        echo "OCP didn't started in $OCP_BOOT_TIMEOUT secs, exit"
-        exit 1
-    fi
-    sleep 2
-    ELAPSED=$((ELAPSED+1))
-  done
+   export PATH=${PATH}:${OCP_TOOLS_DIR}
 }
 
 add_user_as_admin() {
@@ -192,26 +156,24 @@ wait_for_automation_service_broker() {
         $OC_BINARY rollout cancel dc/openshift-automation-service-broker -n=openshift-automation-service-broker
         sleep 5
         $OC_BINARY rollout latest dc/openshift-automation-service-broker -n=openshift-automation-service-broker
-        available=$(${OC_BINARY} get dc/openshift-automation-service-broker -n=openshift-automation-service-broker -o=jsonpath={.status.conditions[0].status})
-        progressing=$(${OC_BINARY} get dc/openshift-automation-service-broker -n=openshift-automation-service-broker -o=jsonpath={.status.conditions[1].status})
-        DEPLOYMENT_TIMEOUT_SEC=1200
+        DESIRED_REPLICA_COUNT=1
+        CURRENT_REPLICA_COUNT=$(${OC_BINARY} get dc/openshift-automation-service-broker -n=openshift-automation-service-broker -o=jsonpath='{.status.availableReplicas}')
+        DEPLOYMENT_TIMEOUT_SEC=300
         POLLING_INTERVAL_SEC=5
         end=$((SECONDS+DEPLOYMENT_TIMEOUT_SEC))
-        while [[ "${available}" != "\"True\"" || "${progressing}" != "\"True\"" ]] && [ ${SECONDS} -lt ${end} ]; do
-            available=$(${OC_BINARY} get dc/openshift-automation-service-broker -n=openshift-automation-service-broker -o json | jq ".status.conditions[] | select(.type == \"Available\") | .status")
-            progressing=$(${OC_BINARY} get dc/openshift-automation-service-broker -n=openshift-automation-service-broker -o json | jq ".status.conditions[] | select(.type == \"Progressing\") | .status")
-            timeout_in=$((end-SECONDS))
-            echo "Deployment is in progress...(Available.status=${available}, Progressing.status=${progressing}, ${timeout_in} seconds remain)"
-            sleep ${POLLING_INTERVAL_SEC}
+        while [ "${CURRENT_REPLICA_COUNT}" -ne "${DESIRED_REPLICA_COUNT}" ] && [ ${SECONDS} -lt ${end} ]; do
+          CURRENT_REPLICA_COUNT=$(${OC_BINARY} get dc/openshift-automation-service-broker -n=openshift-automation-service-broker -o=jsonpath='{.status.availableReplicas}')
+          timeout_in=$((end-SECONDS))
+          echo "Deployment is in progress...(Current replica count=${CURRENT_REPLICA_COUNT}, ${timeout_in} seconds remain)"
+          sleep ${POLLING_INTERVAL_SEC}
         done
-        if [ "${progressing}" == "\"True\"" ]; then
-            echo "Ansible service catalog deployed successfully"
-        elif [ "${progressing}" == "False" ]; then
-            echo "Ansible service catalog deployment failed. Aborting."
-            exit 1
+
+        if [ "${CURRENT_REPLICA_COUNT}" -ne "${DESIRED_REPLICA_COUNT}"  ]; then
+          printError "Automation Service Broker deployment failed. Aborting"
+          exit 1
         elif [ ${SECONDS} -ge ${end} ]; then
-            echo "Deployment timeout. Aborting."
-            exit 1
+          printError "Deployment timeout. Aborting."
+          exit 1
         fi
     fi
 }
@@ -224,7 +186,6 @@ run_ocp() {
     fi
     $OC_BINARY cluster up --public-hostname="${OC_PUBLIC_HOSTNAME}" \
                           --routing-suffix="${OC_PUBLIC_IP}.${DNS_PROVIDER}"
-    wait_ocp
     wait_for_automation_service_broker
 }
